Fix PPU bugs: VBlank flag, palette reads, A12 tracking, sprites
Phase 1 PPU bug fixes:
- Bug #1: set_vblank_flag() no longer called every VBlank dot
- Bug #2: Palette PPUDATA reads now always increment VRAM address
- Bug #3: Remove phantom tick_single_dot() on PPUDATA access
- Bug #4: Correct palette mirror behavior at
- Bug #13: Convert rendering disable static local to member variable

Additional PPU improvements:
- Hardware-accurate A12 tracking for MMC3 IRQ (track_a12_line)
- Cycle-accurate sprite evaluation with secondary OAM
- Background shift register pipeline with fine X scrolling
- Pre-render scanline properly clears sprite flags at dot 1

+++ tests/ppu/vram_address_tests.cpp
@@ -151,7 +151,7 @@ TEST_CASE_METHOD(VRAMAddressTestFixture, "VRAM Address Wrapping", "[ppu][vram][w
 
 		// Write to base nametable
 		set_vram_address(0x2000);
-		write_ppu_register(0x2007, test_data);
+		write_ppu_register(0x2007, static_cast<uint8_t>(test_data));
 
 		// Check mirrors
 		for (uint16_t mirror = 0x3000; mirror <= 0x3F00; mirror += 0x1000) {
